refactor: member table to _user table
