+++ app/src/main/frontend/src/pages/Signup.jsx
@@ -1,6 +1,6 @@
 import MainContainer from "../components/MainContainer";
 import Navbar from "../components/Navbar";
-import SignupForm from "../components/login/SignupForm";
+import SignupForm from "../components/form/SignupForm";
 const Signup = () => {
   return (
     <>
