--- app/src/main/java/com/written/app/hello/HelloController.java
@@ -1,19 +1,0 @@
-package com.written.app.hello;
-
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@RestController
-public class HelloController {
-
-    @GetMapping("/api/test")
-    public Map<String, String> hello() {
-        Map<String, String> response = new HashMap<>();
-        response.put("message", "hello world!!!!");
-        return response;
-    }
-}
