--- app/src/main/java/com/written/app/member/dto/MemberFormDto.java
@@ -1,10 +1,0 @@
-package com.written.app.member.dto;
-
-import lombok.Data;
-
-@Data
-public class MemberFormDto {
-    private String id;
-    private String pwd;
-    private String confirmPwd;
-}
