--- app/src/main/java/com/written/app/member/vo/MemberVo.java
@@ -1,13 +1,0 @@
-package com.written.app.member.vo;
-
-import lombok.Data;
-
-@Data
-public class MemberVo {
-    private String memberNo;
-    private String id;
-    private String pwd;
-    private String nick;
-    private String delYn;
-    private String createdDate;
-}
